added a test
